Several improvements to many of the files
Added markdown files, modified the ps script by adding the growth rate to the individual ulva and hypnea data frames so I can run regressions with ps vs growth.

Added some ggplots to the markdown that may or may not work at this point. They work in the .R file.